backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-multipart==0.0.9
rdflib==7.0.0
pandas==2.2.2
openpyxl==3.1.3
httpx==0.27.0
python-dotenv==1.0.1
